requirements.txt
langchain
python-dotenv
streamlit
faiss-cpu
openai
google-search-results
pypdf
chromadb
tiktoken
